Configure relabel engine for 1024-L1 teacher NNUE

+++ Fairy-Stockfish/src/uci.cpp
@@ -141,9 +141,9 @@ namespace {
       PackedSfen sfen;
       std::int16_t score;
       std::uint16_t move;
-      std::uint16_t gamePly;
+      std::int8_t gamePly;
       std::int8_t game_result;
-      std::uint8_t padding;
+      std::uint16_t padding;
     };
 
     static_assert(sizeof(PackedSfenValue) == PackedSfenDataBits / 8 + 8,
